--- src/main/java/frc/team449/generalInterfaces/SlaveMotor.java
@@ -1,4 +1,0 @@
-package frc.team449.generalInterfaces;
-
-/** A marker interface for followers */
-public interface SlaveMotor {}
